// js/src/types/StatementSchema.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RecordMapping } from "./RecordMapping";

export type StatementSchema = { id: number, name: string, record_mapping: RecordMapping, };

export type StatementSchemaFields = { name: string, record_mapping: RecordMapping, };

export type StatementSchemas = { schemas: Array<StatementSchema>, };